Fixed Blog section
Removed the <img> and made the image a background for the .blog div instead. Removed relative positioning & added margins instead. For this case it's easier to position and more predictable for mobile & tablet devices. Moved .blogButtons into the .blogText div so that it will always be inline with that text. Added and opaque background the .blogText on mobile so that it can be read against the busy image.